% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.fabOF.R
\name{print.fabOF}
\alias{print.fabOF}
\title{Print fabOF}
\usage{
\method{print}{fabOF}(x, ...)
}
\arguments{
\item{x}{Object of class fabOF}

\item{...}{Further arguments passed to or from other methods.}
}
\description{
Prints description of fabOF model object.
}
\author{
Philip Buczak
}
