% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.fabOF.R
\name{predict.fabOF}
\alias{predict.fabOF}
\title{Predict fabOF}
\usage{
\method{predict}{fabOF}(object, newdata, ...)
}
\arguments{
\item{object}{Fitted fabOF model object.}

\item{newdata}{Dataset containing new observations to be predicted.}
}
\value{
Predicted ordinal response category labels.
}
\description{
Predicts ordinal response categories for new observations using a fabOF model fit.
}
\author{
Philip Buczak
}
